server/requirements.txt
flask
flask-cors
torch
openai-whisper
pydub
pyaudio
numpy==1.24.1
